move more stuff to oldviews
